+++ src/app/recruit/[id]/edit/page.tsx
@@ -1,52 +1,36 @@
 'use client'
 
-import { useEffect, useState } from 'react'
-import { useRouter } from 'next/navigation'
-import useToast from '@/hook/useToast'
 import { Box, Button, Stack, TextField, Typography } from '@mui/material'
-import Image from 'next/image'
+import RowRadioButtonsGroup from './panel/radioGroup'
+import SetTeamRole from './panel/SetTeamRole/SetTeamRole'
+import TagAutoComplete from './panel/SetTeamTag/TagAutoComplete'
+import { useEffect, useState } from 'react'
+import BasicSelect, { ComponentType } from './panel/BasicSelect'
+import SetInterview from './panel/SetInterview/SetInterview'
+import SetCommunicationToolLink from './panel/SetCommunicationToolLink/SetCommunicationToolLink'
+import useToast from '@/hook/useToast'
+import SelectRegion from './panel/SelectRegion'
 import ImageUploadButton from '@/components/ImageUploadButton'
-import RowRadioButtonsGroup from '../[id]/edit/panel/radioGroup'
-import SetTeamRole from '../[id]/edit/panel/SetTeamRole/SetTeamRole'
-import TagAutoComplete from '../[id]/edit/panel/SetTeamTag/TagAutoComplete'
-import BasicSelect, { ComponentType } from '../[id]/edit/panel/BasicSelect'
-import SetInterview from '../[id]/edit/panel/SetInterview/SetInterview'
-import SetCommunicationToolLink from '../[id]/edit/panel/SetCommunicationToolLink/SetCommunicationToolLink'
-import SelectRegion from '../[id]/edit/panel/SelectRegion'
-import { ITag } from '@/types/IPostDetail'
-import useAxiosWithAuth from '@/api/config'
+import Image from 'next/image'
 import useSWR from 'swr'
-// import useAuthStore from '@/states/useAuthStore'
-
-const dummyData1: ITag = {
-  name: 'java',
-  color: 'red',
-}
-const dummyData2: ITag = {
-  name: 'spring',
-  color: 'blue',
-}
-const dummyData3: ITag = {
-  name: 'react',
-  color: 'green',
-}
-const dummyDatas: ITag[] = [dummyData1, dummyData2, dummyData3]
+import useAxiosWithAuth from '@/api/config'
+import { useRouter } from 'next/navigation'
+import RowRadioButtonsGroupStatus from './panel/radioGroupStatus'
+import { ITag, statusEnum } from '@/types/IPostDetail'
 
 export interface IRoleData {
-  // types 로 병합예정
   role: string | null
   member: number
 }
 
 export interface IFormInterview {
-  // 수정 및 types로 병합예정
   question: string
   type: string
   optionList?: string[]
   ratioList?: { max: string; valueOfMin: string; valueOfMax: string }
 }
 
-const CreateTeam = () => {
+const CreateTeam = ({ params }: { params: { recruit_id: string } }) => {
   const [title, setTitle] = useState<string>('')
   const [image, setImage] = useState<File[]>([])
   const [previewImage, setPreviewImage] = useState<string>(
@@ -63,37 +47,39 @@ const CreateTeam = () => {
   const [content, setContent] = useState<string>('')
   const [roleList, setRoleList] = useState<IRoleData[]>([])
   const [interviewList, setInterviewList] = useState<IFormInterview[]>([])
-  const [allTagList, setAllTagList] = useState<ITag[]>(dummyDatas)
   const [openBasicModal, setOpenBasicModal] = useState(false)
+  const [status, setStatus] = useState<statusEnum>(statusEnum.BEFORE)
+  const [allTagList, setAllTagList] = useState<ITag[]>([])
   const { CuToast, isOpen, openToast, closeToast } = useToast()
   const [toastMessage, setToastMessage] = useState<string>('')
   const router = useRouter()
   const axiosInstance = useAxiosWithAuth()
-  // const { isLogin } = useAuthStore()
-
-  // 로그인 하지 않고 모집글쓰기 들어왓을때 로그인 하러가기 안내가 필요합니다.
-  // 새로운 공통컴포넌트 모달창을 만들면 효율적일듯합니다.
-  // useEffect(() => {
-  //   if (!isLogin) {
-  //     setToastMessage('로그인이 필요합니다')
-  //     openToast()
-  //     router.push('/')
-  //   }
-  // }, [isLogin])
-
   const { data, error } = useSWR(
-    `${process.env.NEXT_PUBLIC_API_URL}/api/v1/recruit/write`,
-    (url: string) => axiosInstance.get(url).then((res) => res.data),
+      `${process.env.NEXT_PUBLIC_API_URL}/api/v1/recruit/edit/${params.recruit_id}`,
+    (url:string) => axiosInstance.get(url).then((res) => res.data),
   )
 
   useEffect(() => {
     if (error) {
       console.log('error ocurred!!')
-      setToastMessage('태그를 불러오는데 실패했습니다.')
+      setToastMessage('데이터를 불러오는데 실패했습니다.')
       openToast()
-    } else if (data) {
-      console.log('tag fetching success', data)
-      setAllTagList(data)
+    } else {
+      setTitle(data.title)
+      setPreviewImage(data.previewImage)
+      setName(data.name)
+      setType(data.type)
+      setTagList(data.tagList)
+      setRegion(data.region)
+      setPlace(data.place)
+      setMonth(data.due)
+      setTeamsize(data.teamsize)
+      setCommunicationTool(data.link)
+      setContent(data.content)
+      setRoleList(data.roleList)
+      setInterviewList(data.interviewList)
+      setStatus(data.status)
+      setAllTagList(data.tagList)
     }
   }, [data])
 
@@ -103,7 +89,7 @@ const CreateTeam = () => {
     }
     try {
       const response = await axiosInstance.post(
-        `${process.env.NEXT_PUBLIC_API_URL}/api/v1/recruit/write`,
+        `${process.env.NEXT_PUBLIC_API_URL}/api/v1/recruit/edit/${params.recruit_id}`,
         {
           place,
           image,
@@ -117,9 +103,10 @@ const CreateTeam = () => {
           tagList,
           roleList,
           interviewList,
+          status,
         },
       )
-      router.push(`/recruitment/${response.data}`) // 백엔드에서 리턴값으로 새로생긴 모집글의 id 를 던져줌
+      if (response.status === 200) router.push(`/recruit/${response.data}`) // 백엔드에서 리턴값으로 새로생긴 모집글의 id 를 던져줌
     } catch (error) {
       console.log(error)
       setToastMessage('모집글 작성 실패, 다시 시도해주세요')
@@ -161,6 +148,10 @@ const CreateTeam = () => {
             value={name}
             onChange={(e) => setName(e.target.value)}
           />
+        </Box>
+        <Box>
+          <Typography>팀 진행 상태</Typography>
+          <RowRadioButtonsGroupStatus setValue={setStatus} />
         </Box>
         <Box>
           <Typography variant="h6">팀 분류</Typography>
@@ -216,13 +207,11 @@ const CreateTeam = () => {
         </Stack>
         <Box>
           <Typography variant="h6">태그</Typography>
-          {allTagList ? (
-            <TagAutoComplete
-              datas={tagList}
-              setData={setTagList}
-              allTagList={allTagList}
-            />
-          ) : null}
+          <TagAutoComplete
+            datas={tagList}
+            setData={setTagList}
+            allTagList={allTagList}
+          />
         </Box>
         {type === 'study' ? null : (
           <Box>
