+++ src/app/recruit/[id]/page.tsx
